Update guide for accuracy and use of Docker (#19)
* Update guide for accuracy and use of Docker

This commit updates multiple aspects of the getting started guide.
Most notably, the learner is now instructed to use Docker to
start the external Vault server. A new Docker Compose file is provided
in the repository and instructions are given how to run it.

Additional changes were made to the guide to facilitate a clean
compile. One change was to remove the failing test in the complete
folder. This test was dependent on an externally running container.
This test was removed. Additionally, a main class was added to the
initial folder so that build commands could be run from a CI system.

The core teachings and code were not impacted by this PR.

* Add polish to the README and minor build config

Update language used in the README per PR review.
Modified the usage of dependency-management plugin in
Gradle build files to align with Boot recommendations.
Updated Gradle wrapper in complete folder.